+++ cloudpickle/cloudpickle.py
@@ -78,9 +78,6 @@ else:
     PY3 = True
 
 
-# caches dynamic modules that are not referenced in sys.modules
-_dynamic_modules_globals = {}
-
 
 def _make_cell_set_template_code():
     """Get the Python compiler to emit LOAD_FAST(arg); STORE_DEREF
@@ -1094,14 +1091,12 @@ def _make_skel_func(code, cell_count, base_globals=None):
     if base_globals is None:
         base_globals = {}
     elif isinstance(base_globals, str):
-        base_globals_name = base_globals
-        if base_globals_name in sys.modules.keys():
+        if sys.modules.get(base_globals, None) is not None:
             # this checks if we can import the previous environment the object
             # lived in
-            base_globals = vars(sys.modules[base_globals_name])
+            base_globals = vars(sys.modules[base_globals])
         else:
-            base_globals = _dynamic_modules_globals.get(base_globals_name, {})
-            _dynamic_modules_globals[base_globals_name] = base_globals
+            base_globals = {}
 
     base_globals['__builtins__'] = __builtins__